handled delete category with product

+++ src/main/java/com/mycompany/pr/farmacia/Servlets/SvDeleteCategory.java
@@ -5,9 +5,13 @@
 package com.mycompany.pr.farmacia.Servlets;
 
 import com.mycompany.pr.farmacia.Controllers.CategoryController;
+import com.mycompany.pr.farmacia.Controllers.ProductController;
+import com.mycompany.pr.farmacia.Entities.Category;
+import com.mycompany.pr.farmacia.Entities.Product;
 import com.mycompany.pr.farmacia.Persistence.exceptions.NonexistentEntityException;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.ServletException;
@@ -15,6 +19,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  *
@@ -24,6 +29,9 @@ import javax.servlet.http.HttpServletResponse;
 public class SvDeleteCategory extends HttpServlet {
 
     CategoryController categoryController = new CategoryController();
+    ProductController productController = new ProductController();
+
+    List<Product> productsList = productController.getProductsController();
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
@@ -36,14 +44,35 @@ public class SvDeleteCategory extends HttpServlet {
             throws ServletException, IOException {
 
         int id = Integer.parseInt(request.getParameter("id"));
+        Category category = categoryController.getCategoryController(id);
 
-        try {
-            categoryController.deleteCategoryController(id);
-        } catch (NonexistentEntityException ex) {
-            Logger.getLogger(SvDeleteCategory.class.getName()).log(Level.SEVERE, null, ex);
+        boolean categoryInUse = false;
+
+        int amountProductsWithCategory = 0;
+
+        for (Product product : productsList) {
+            if (product.getCategory().equals(category.getName())) {
+                categoryInUse = true;
+                amountProductsWithCategory++;
+                break;
+            }
         }
 
-        response.sendRedirect("Categories.jsp");
+        HttpSession httpSession = request.getSession();
+
+        httpSession.setAttribute("amountProductsWithCategory", amountProductsWithCategory);
+
+        if (categoryInUse) {
+            response.sendRedirect("CategoryInUse.jsp");
+        } else {
+            try {
+                categoryController.deleteCategoryController(id);
+                response.sendRedirect("Categories.jsp");
+                System.out.println("Categoría eliminada exitosamente.");
+            } catch (NonexistentEntityException ex) {
+                Logger.getLogger(SvDeleteCategory.class.getName()).log(Level.SEVERE, null, ex);
+            }
+        }
 
         processRequest(request, response);
     }
